Make length incrementing/decrementing lock-free.
An interesting consequence of this is that without as many
contended futex syscalls, the kernel can't get smart about
scheduling other threads when a given thread is just waiting for
the queue to not be empty or full. That means that our whole
"spin until it works" thing in the test becomes DRAMATICALLY less
efficient. Incidentally, this is why I decreased the iteration
count.

Just one of the many strange things about concurrent
programming... :)