Implement a preliminary MPSC lock-free queue.
First of all, this is an implementation for a multi-producer,
single-consumer queue, which will eventually be used internally
to implement a full MPMC queue.

I chose this design because using subqueues makes things easier
to implement, as well as this particular way of doing it makes
it easy to do the AOS thing where multiple consumers are all
guaranteed to see all the messages. (Though I don't think this is
technically standard behavior, I find it EXTREMELY useful.)