Add memory fencing and split constants out.
I know it was working initially, but it was technically not
guaranteed to work always, because the CPU could have reordered
instructions and broken things. Now that should no longer be an
issue.

I also made a bunch more things volatile, which should guard
against compiler reordering as well.

As far as constants go, I figured it was better to just have them
all in one place for easy modification.