Finally get memory management in order.
There are now facilities that can safely and easily free shared
memory. Everything is manually managed, the rational being that
this is something the average user would probably use rarely if
at all. (Case in point: I don't believe the original AOS
implementation even HAD a facility for freeing and reallocating
shared memory.) Given that, it's probably best to let people do
it manually in order to avoid the rare but possible
hard-to-find bugs that reference counting or something similar
would open the door to.