Fix issue with blocking and normal ops.
When we use blocking writes with non-blocking reads, there is
no code in the reading function to actually go and wake the
writer if it has blocked, leading to deadlocks. I'm not sure if
this was a bug, or a deliberate (if undocumented) constraint in
order to ensure that the non-blocking dequeue operation never
enters the kernel. Either way, I've now gone through and fixed
it properly, such that the non-blocking dequeue is still
guaranteed to remain in userspace provided it's not being used on
the same queue with blocking enqueue operations.